sync-async-service-pipeline/pyproject.toml
[tool.poetry]
name = "zeromq-patterns"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyzmq = "^25.1.2"
requests = "^2.31.0"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
